generator.py: Drop toint() of integer value
The recently introduced toint() function does not accept arguments
of type int and will trigger an exception in this case.
This leads to an always crashing code path in generator.py,
where rm_dict[] maps the rounding mode string to an integer,
that is then passed to toint().

There are several ways to address this (e.g. mapping to a string,
or making toint() more robust), but dropping the call to toint()
is obviously the right thing to do.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>